% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_item.R
\name{ez_item}
\alias{ez_item}
\title{Item Analysis}
\usage{
ez_item(data, method = "auto")
}
\arguments{
\item{data}{Data of item}

\item{method}{Method like "2PL", "graded", and "auto"}
}
\value{
A character vector.
}
\description{
Item Analysis
}
\examples{
data <- survey[5:10]
strsplit(data)
}
