% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_item_plot.R
\name{ez_item_plot}
\alias{ez_item_plot}
\title{Item Analysis}
\usage{
ez_item_plot(data)
}
\arguments{
\item{data}{Data of item}

\item{type}{Type like "auto", "CTT", "IRT"}
}
\value{
A graph
}
\description{
Item Analysis
}
\examples{
data <- survey[5:9]
ez_item_plot(data)
}
