% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezoutput-package.R
\docType{package}
\name{ezoutput-package}
\alias{ezoutput}
\alias{ezoutput-package}
\title{ezoutput: Easy Statistics Output}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jongheon Kim \email{cysics@daum.net}

}
\keyword{internal}
