% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\docType{data}
\name{survey}
\alias{survey}
\title{Survey about Tteokbokki}
\format{
Tibble with 335 rows and 57 variables:
\describe{
\item{gender}{gender}
\item{age}{age}
\item{grade}{grade}
}
}
\source{
<박민균>
}
\usage{
survey
}
\description{
Evaluation of the store
}
\examples{
  survey
}
\keyword{datasets}
