% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_item.R
\name{ez_item}
\alias{ez_item}
\title{Item Analysis}
\usage{
ez_item(data, method = "auto", k = 3, l = 1, u = 3, csv = FALSE)
}
\arguments{
\item{data}{Data of item}

\item{method}{Method like "2PL", "graded", and "auto"}

\item{k}{Number of group}

\item{l}{Lower of group}

\item{u}{Upper of group}
}
\value{
A data.frame.
}
\description{
Item Analysis
}
\examples{
data <- survey[5:9]
ez_item(data)
}
