% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\docType{data}
\name{survey}
\alias{survey}
\title{Survey about Tteokbokki}
\format{
Tibble with columns
\describe{
\item{성별}{gender}
\item{연령}{age}
\item{학년}{grade}
}
}
\source{
<박민균>
}
\usage{
survey
}
\description{
Evaluation of the store
}
\examples{
  survey
}
\keyword{datasets}
